+++ lib/responsiveness.dart
@@ -1,5 +1,6 @@
 ///a responsive framework using the bootstrap breakpoints
 library responsiveness;
 
-export 'src/screen_size/screen_size.dart';
+export 'src/screen_size.dart';
 export 'src/responsive_widgets/responsive_widgets.dart';
+