+++ lib/src/generators/generated_file_warning_generator.dart
@@ -1,0 +1,9 @@
+import 'generator.dart';
+import 'screen_sizes.dart';
+
+class GeneratedFileWarningGenerator implements Generator{
+  @override
+  String generateSourceCode(ScreenSizes screenSizes) {
+    return  "// GENERATED CODE - DO NOT MODIFY BY HAND\n";
+  }
+}