+++ lib/src/generators/generator.dart
@@ -1,0 +1,5 @@
+import 'screen_sizes.dart';
+
+abstract class Generator{
+  String generateSourceCode(ScreenSizes screenSizes);
+}