+++ lib/src/generators/selector_generator.dart
@@ -1,7 +1,12 @@
-import 'package:flutter/material.dart' show BuildContext;
-import '../screen_size.dart';
+import 'screen_sizes.dart';
 
-extension Selector on BuildContext{
+import "generator.dart";
+
+class SelectorGenerator implements Generator {
+  @override
+  String generateSourceCode(ScreenSizes screenSizes) {
+    return """
+extension _Selector on BuildContext{
   T select<T>(List<T?> values){
     assert(values.length == ScreenSize.values.length, "The length of values list has to be equal to the number of different screen sizes");
     assert(values[0] != null, "The first element in the values list can not be null");
@@ -12,5 +17,8 @@ extension Selector on BuildContext{
       }
     }
     throw Exception("Error: It was not possible to determine a value for the current screen size");
+  }
+}
+""";
   }
 }