+++ CHANGELOG.md
@@ -78,3 +78,13 @@
 ## 1.1.8
 
 - fixed typo in README
+
+## 1.1.9(includes breaking changes)
+
+- got rid of ```ScreenSizeProvider``` (breaking change)
+- got rid of ```MinimumWidths``` class (breaking change)
+- ```ScreenSize``` class is now an enum (breaking change)
+- added code generation for providing custom breakpoints
+
+
+